// types/variationsOptions.tsx
export enum variationsOptions {
  Add = "Add variation",
  CreateAll = "Create variations from all attributes",
  DeleteAll = "Delete all variations",
}

export enum variationsPriceOptions {
  SetRegular = "Set regular price",
  IncreaseRegular = "Increase regular price",
  DecreaseRegular = "Decrease regular price",
  SetSales = "Set sales price",
  IncreaseSales = "Increase sales price",
  DecreaseSales = "Decrease sales price",
  SetScheduled = "Set scheduled sales date",
}

export enum variationsStockOptions {
  SetStockQty = "Set stock quantity",
  SetInStock = "Set status - in stock",
  SetOutOfStock = "Set status - out of stock",
}